abstract class for scraper classes for websites
