+++ MERG_PSI/WebScraper/KampasScraper.cs
@@ -5,19 +5,19 @@ using System.Threading.Tasks;
 
 namespace WebScraper
 {
-    public class KampasScraper
+    public class KampasScraper : SiteScraper
     {
-        private const string _websiteLink = @"https://www.kampas.lt";
-        private const string _subdirectory = @"/butai";
-        private const string _pageQueryString = @"page=";
         public List<RealEstate> ScrapedRealEstate { get; set; } = new List<RealEstate>();
         private bool _reachedPageNoAds = false;
 
         public event EventHandler<ScrapingDomainEventArgs> ScrapingDomain;
 
-        public KampasScraper(Form1 myUI)
+        public KampasScraper(Form1 myUI, string websiteLink, string subdirectory, string pageQueryString)
         {
             ScrapingDomain += myUI.OnScrapingDomain;
+            _websiteLink = websiteLink;
+            _subdirectory = subdirectory;
+            _pageQueryString = pageQueryString;
         }
 
         public async Task ScrapeKampasWebsite()
@@ -56,36 +56,6 @@ namespace WebScraper
                 }
                 else { _reachedPageNoAds = true; }
             }
-        }
-
-        private bool IsAdHasAllNeededData(string link, string mapLink, int numberOfRooms, double scrapedPrice, double pricePerSqM, double area, string mapCoords)
-        {
-            var calculatedPrice = pricePerSqM * area;
-            if (
-                //fix put area and priceerarea
-
-                mapCoords == "" ||
-                link == "" ||
-                mapLink == "" ||
-                numberOfRooms == 0 ||
-                (calculatedPrice == 0 && scrapedPrice == 0) ||
-                !IsValuesClose(calculatedPrice, scrapedPrice, 1000)
-                /*Municipality == /*fix to be implemented*/
-                /*Street == /*fix to be implemented*/)
-            {
-                return false;
-            }
-            else { return true; }
-        }
-
-        public bool IsValuesClose(double value1, double value2, int roundErr)
-        {
-            var diff = value1 - value2;
-            if ((diff < roundErr && diff >= 0) || (diff <= 0 && diff > -roundErr))
-            {
-                return true;
-            }
-            else { return false; }
         }
     }
 }
