+++ MERG_PSI/WebScraper/N9Scraper.cs
@@ -1,0 +1,7 @@
+﻿namespace WebScraper
+{
+    public class N9Scraper
+    {
+
+    }
+}
